Refactor code
This commit rewrites the C++ bits in C, and adjusts the architecture by
trying to decouple classes. Additionally, this removes the verbose
command-line option (as that really only set the G_MESSAGES_DEBUG
environment variable) and adds --system-bus to allow running on the
system bus. That is needed for Cockpit, since it allows managing the
root user, and there isn’t a session bus to connect to.

Signed-off-by: Ernestas Kulik <ekulik@redhat.com>