general: Fix memory management for daemon and service
Neither ReportdDaemon nor ReportdService were being properly freed when
quitting, as the service D-Bus object is not being unexported, and the
internal references get caught in a cycle. This commit shuffles code
around to ensure that the daemon destroys the service properly, in turn
ensuring that the service drops the reference to the daemon.

Signed-off-by: Ernestas Kulik <ekulik@redhat.com>